separated docs. no errors, just split up db, store, routes
