+++ server/DB/conn.js
@@ -1,0 +1,3 @@
+const Sequelize = require('sequelize');
+const conn = new Sequelize(process.env.DATABASE_URL || 'postgres://localhost/playback');
+module.exports = conn;
