+++ src/Cart.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { getProducts, setCartThunks, destroyCartThunks, updateThunks } from '../store';
+import { getProducts, setCartThunks, destroyCartThunks, updateThunks } from './redux/store.js';
 
 class _Cart extends React.Component {
   constructor() {
@@ -22,7 +22,6 @@ class _Cart extends React.Component {
     await this.props.setCart();
   }
   render() {
-    console.log(this.props.cart);
     const { products, cart, destroy, update } = this.props
     const items = cart.map( c => {
       return {...c, cart: products.filter(product => c.productId === product.id)};
@@ -37,8 +36,8 @@ class _Cart extends React.Component {
                 <Link to={`/products/${item.id}`} activeclassname="active"><li key='img'><img src ={item.product.imageURL}></img></li></Link>
                 <Link to={`/products/${item.id}`} activeclassname="active"><h1 key='name'>{item.product.name}</h1></Link>
                <li key='genre'>{item.product.genre}</li>
-               <li key='quantity'>{item.quantity}</li>
-                <li key='price'>${item.product.price}</li>
+               <li key='quantity'>Quantity: {item.quantity}</li>
+                <li key='price'>Cost: ${item.product.price}</li>
 
                 <button onClick = { () => update(item.id, 'add') } >+</button>
                 <button onClick = { () => update(item.id, 'subtract') } >-</button>
